requirements.txt
Django
django-registration
django-widget-tweaks
djangorestframework
django-filter
django-smart-selects
django_excel
mysqlclient
django-mysql
setuptools
gunicorn
dj-database-url
psycopg2